Update indexOfAll snippet return type
Now we're just going to return an empty array. This keeps a consistent return type, does not need checking, and won't break api chaining.

Decision was made because otherwise you'd have to add a filter step or check if the return is not an array. Instead we just want an empty array so it has a consistent API, and no indeces is a valid amount