update buttons and nav semantics
the back to top button and menu button each lack appropriate accessible names.

The primary page navigation is presently contained within an `aside` rather than a `nav` element.

This commit adds accessible names to the previously mentioned buttons, so they will be appropriately conveyed to screen readers.  It also changes the navigation from an `aside` to `nav` and provides the navigation with an `aria-label=“table of contents”` as all the links are in page links to the different sections of this document.