Prevent error flash when JS is enabled
When JS is enabled, there's a very brief moment
when the content web components haven't loaded
just yet, so the user sees a flash of the error
message. To prevent this, we can use a `noscript`
tag, along with the styles added into it, so that
the error message will only appear for users with
JS disabled completely.